+++ src/aat/java/uk/gov/hmcts/cmc/claimstore/tests/BaseTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.claimstore.aat.tests;
+package uk.gov.hmcts.cmc.claimstore.tests;
 
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,7 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.cmc.claimstore.aat.Bootstrap;
+import uk.gov.hmcts.cmc.claimstore.tests.Bootstrap;
 import uk.gov.hmcts.cmc.claimstore.processors.JsonMapper;
 
 @RunWith(SpringRunner.class)
